test: 대용량 패킷(50KB x 3) 통째 전송 및 패킷 슬라이싱 검증
50,000바이트 크기의 대형 패킷 3개를 연속 전송하여 수신 및 조립 성공
TCP 스트림 내에서 개별 패킷 경계를 정확히 분리하는 로직 검증
대용량 데이터 수신 시 버퍼 관리 및 메모리 안정성 확인

+++ Network/Common/Protocol.hpp
@@ -12,7 +12,7 @@ enum class S2C_PACKET_ID : uint8_t { kNone, kMove, kChat, kCnt };
 #pragma pack(push, 1)
 
 struct PACKET_HEADER {
-	uint16_t size;
+	int size;
 	uint16_t id;
 };
